requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
paho-mqtt==2.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
typing==3.7.4.3
typing_extensions==4.12.2
Werkzeug==3.1.3
